requirements.txt
agate==1.6.1
asn1crypto==1.4.0
attrs==21.2.0
azure-common==1.1.27
azure-core==1.14.0
azure-storage-blob==12.8.1
Babel==2.9.1
beautifulsoup4==4.9.3
boto3==1.15.18
botocore==1.18.18
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
colorama==0.4.3
cryptography==3.4.7
dbt==0.19.1
dbt-bigquery==0.19.1
dbt-core==0.19.1
dbt-postgres==0.19.1
dbt-redshift==0.19.1
dbt-snowflake==0.19.1
decorator==4.4.2
google-api-core==1.23.0
google-auth==1.30.1
google-cloud-bigquery==2.3.1
google-cloud-core==1.4.4
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.52.0
greenlet==1.1.0
hologram==0.0.13
idna==2.9
importlib-metadata==4.2.0
isodate==0.6.0
jeepney==0.6.0
Jinja2==2.11.2
jmespath==0.10.0
json-rpc==1.13.0
jsonschema==3.1.1
keyring==21.8.0
leather==0.3.3
Logbook==1.5.3
MarkupSafe==2.0.1
mashumaro==2.0
minimal-snowplow-tracker==0.0.2
msgpack==1.0.2
msrest==0.6.21
networkx==2.5.1
numpy==1.20.1
oauthlib==3.1.0
oscrypto==1.2.1
pandas==1.2.3
parsedatetime==2.6
proto-plus==1.18.1
protobuf==3.17.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.10.1
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyrsistent==0.17.3
python-craigslist==1.1.4
python-dateutil==2.8.1
python-slugify==5.0.2
pytimeparse==1.1.8
pytz==2020.5
PyYAML==5.4.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.3.7
SecretStorage==3.3.1
six==1.15.0
snowflake-connector-python==2.3.6
soupsieve==2.2
SQLAlchemy==1.4.15
sqlparse==0.3.1
text-unidecode==1.3
typing-extensions==3.7.4.3
urllib3==1.25.11
Werkzeug==1.0.1
zipp==3.4.1
